+++ src/PhpSpreadsheet/Writer/Xlsx/Drawing.php
@@ -569,6 +569,12 @@ class Drawing extends WriterPart
 
                 $iterator->next();
             }
+            $iterator = $spreadsheet->getSheet($i)->getInCellDrawingCollection()->getIterator();
+            while ($iterator->valid()) {
+                $aDrawings[] = $iterator->current();
+
+                $iterator->next();
+            }
         }
 
         return $aDrawings;